[REFACTOR]: change import lint and thunk logic isolation
