+++ shoppingApp/src/hooks/useBookmark.tsx
@@ -1,5 +1,6 @@
-import { useDispatch } from 'react-redux';
 import { ProductType, ProductApi } from '../modules/productApi';
+import { useDispatch } from 'react-redux';
+
 import getLocalStorage from '../utils/getLocalStorage';
 
 const useBookmark = () => {
